[WFLY-12859]:  Refactoring code to centralize the configuration of the broker.
